requirements.txt
asgiref==3.3.1
boto3==1.17.97
botocore==1.20.97
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.25.0
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.4
Django==3.1.4
django-allauth==0.44.0
django-debug-toolbar==3.2.1
django-redis==5.0.0
django-storages==1.11.1
djangorestframework==3.12.2
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2020.5
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
